requirements.txt
markdown>=3.5
PyYAML>=6.0
openai>=1.0
